PaymentGateway
Assume payment gateway is up until we find it isn't. Because classes are
reloaded every request in development, it creates a race condition where
PaymentGateway.available? may be nil in the request. Because classes are
cached in test and also eager loaded in production this shouldn't be a problem
in those environments, but it's annoying in development