Use ENV.fetch for secrets
This will blow up on a missing value instead of silently running
insecurely.

I really, really dislike YAML with ERB as used in secrets.yml. I'd
prefer to configure these secrets in the various environment files, but
Rails is quite entangled with the choice of YAML here. secrets.yml only
continues to exist here to load from the environment.

We'll need to move from dotenv-rails to dotenv-deployment at some point
in the future as the way we're currently using dotenv is deprecated.