Jadu::Claim.create submits claims to Jadu.
Jadu::Claim.create is just a wrapper around the work of others
so the specs just assert the relevant methods are called on
the collaborator objects. I don't like testing with a lot of
doubles because it's brittle, but I also don't like duplicating
tests already extant. I have attended to the brittleness risk
by using instance_doubles with check the doubled object responds
to mocked methods.