Add a number of healthcheck components
Adds a number of healthcheck components for:
- Barclaycard
- RabbitMQ
- Sendgrid

All components follow a common interface, exposing a class level
'available?' predicate method.

The RabbitMQ spec is quite awful. It has far too much knowledge
of the inner workings of the class as well as well as its
components. For example, the spec assumes a session is returned
when :new is called on Bunny.