Upgrade SimpleForm
Upgrades SimpleForm from 3.1.0rc2 to 3.20.

A noticeable performance decrease was spotted when rendering views in
the app after updating to Rails from 4.2.0 to 4.2.4 - in most cases this
was increased running times by 4-5x! It appears as though the issues lie with
I18n::MissingTranslation's in SimpleForm - something that has been addressed
in a later version of the gem.

Please see: https://github.com/plataformatec/simple_form/issues/1227 for
more info.

Updating SimpleForm required a public `has_attribute?` method to be introduced
to all form objects.