Fixes rabbitmq timeouts by incr heartbeat to 28sec
Setting heartbeat to 28 seconds for both rails and sneakers workers.

There used to be heartbeat timeouts during high workload and initial
handshake in the rabbitmq logs. This was because sneakers hardcoded
heartbeat default is set to 2 seconds, so when negotiating between
client and server (30 sec server and 2 sec client) the result was 2,
making heartbeat intervals as low as 0.4 seconds. When the heartbeats
were missed the connections were being reset. This manifested to
issues where ActiveJobs were not being enqueued properly into
rabbitmq, for example they were being marked as submitted while never
having made it into the actual rabbitmq queues.

More details on:
https://github.com/ministryofjustice/atet/pull/801