Add payment gateway task logging
Adds logging to the payment gateway when the state changes from
available = true to false and vice versa. This should allow someone
to detect whether the gateway ever becomes unavailable
(spoiler alert: it does become unavailable often).
The debug statement will be removed later, but for now leaving it in
as we 'll need it on staging.