Add pdftk installation instructions to the readme
It's important to note that a new version
of pdftk is required on OS X 10.11, Macs that have been
upgraded & use the old version of pdftk cause the app
to hang on pdf generation.

This commit introduces links to the relevant versions
of pdftk that need to be installed for the application
to function correctly as well as changing some of the
layout of the document.