Stop pdftk version differences breaking spec
The latest version of OS X stopped pdftk from working
correctly - this means that some developer machines might
be running different versions of pdftk & different
versions will produce different file outputs and therefore
different md5 checksums. Seems as we already test the
content that is injected into the pdf in an other spec it
should be safe for us to just assert we send three strings
that look like md5 checksums here.

Ideally this wouldn't be a problem if we ran the app in a
consistent manner when developing.