Update validation copy after PDF/app sync
With the removal of one of the acas exemption reasons
we now have claims that people can return to(i.e not
submitted to jadu yet) that have the old acas reason
associated with one or more of the respondents. This
shouldn't be a problem though as the save and complete
forces a user to go through all of the form pages when
the user returns. As the old reason is no longer in the
the accepted reasons lists the respondents & additional
respondents pages will fail validation, showing the user
an inclusion error message - which is added in this
commit - prompting them to select a new reason.

Claims that have been marked as submitted or payment
required needn't be concerned with this legacy change as
the claim is locked out & besides Jadu would have
already received the claim.