First working version which brings up unicorn
This is the first working version that brings up unicorn. Some additional
files have been carried through that may not be explicitly used. The
rails Dockerfile is the ultimate reference for what is used and what isn't.

To build with tagname "et3" run from the top level of the repo:
  docker build  -t et3 -f docker-ng/rails/Dockerfile .

To run priviledged mode:

  docker run -ti -p 8080:8080 \
    --env-file=<path to>/jenkins-vars-censored.sh-env  \
    -v $PWD/<path to>/jenkins-vars-censored.sh:/etc/envvars \
    et3

where jenkins-vars-censored.sh-env are the docker style env vars and
jenkins-vars-censored.sh are the bash style env vars (ie. can be "source"d)

To run unpriviledged mode, run the above with the following argument:
  /usr/sbin/runsvdir-start