Add submit buttons to admin
We have situations where a claim may need to be manually resubmitted to
jadu, or simply marked as submitted (to silence alerts).

Two buttons have been added to the admin screen - one to resubmit a
claim and another to mark a claim as submitted.  In both instances an event
is recorded against the claim to indicate that the action was triggered
by an admin user.