Feature/rst 1346 translation (#933)
* Moving en translation to its own folder

As a plan to add Welsh translation it's better to tidy up the structure.

* Adding translation files for Welsh

* Adding default url options

* Replacing md files with html slim template

* Adding translation for guides

Refactoring the guides controller so it inherits from application controller.

* Removing guide presenter because it's not needed

* Making rubocop happy

* Fixing failing tests

Adding back the Guide presenter because it's used by other pages. I will remove it after I shift all the md template files to html format. Fixing other bugs by adding or removing locale into a params.

* Terms and conditions are using html template

There is no need to use .md template and it makes the multilingual version more difficult. This is the new T&C page so the copy is not exactly same ans the previous T&Cs.

* Making Rubocop happy after rebase from master

We updated Rails version and all gems that we could. Including rubocop. I fix all the complains execpt the one about allow_any_instance_of. Which I used to fix multiple expect blocks. Those blocks were catching the exception called by the code on purpose. If I would have to remove the allow_any_instance_of too I would have to stub more things which I didn't want to.

* Adding en and cy translation for T&Cs

* Adding language switch link to the header

* Bugfix persistance of language in forms

When you choose the welsh and submit the form then it went directly to english again. I'm not sure if it's beacuse of a Rails 5 issue or I missed something but I fixed it by adding the locale to session.

* Fixing feedback link and adding more translations

* Adding more missing translations

* Fixing more translation issues and css formating

* Another batch of translation and css fixes

* Fix for bug RST-1423-break-line

When you have a multiple files uploaded it displlayed the BR tag as a text not an html.

* Adding translation for js char counter

* Fixing formating

* Fixing missing header for group clame in Welsh

* Error pages are now dynamic

I switch to dynamic error pages so we can easily use the translation.

* Fixing bug RST-1433 for empty current situation

When you choose that you are still employed and do not choose the reason it failed. I add the validation so it displays error message now and you have to select the reason.

* Making rubocop happy

* Pre setting en langage for all tests

* Updating welsh translation for respondent yes/no

In Welsh it depends on context what the translation is.

* More welsh translation updates

* Adding Welsh default values

Days months and other default values were missing so I download them from the I18n repository.

* Fixing bug with unknown locale value

Because we have a dynamic errors and they are using locale when you passed invalid locale value it failed to render correct error page.

* More translation fixes

* Removing test that was removed in master.

* Diversity translation

Adding switch and the translation file.

* Updating email translation

* Updating Access detail plain email with translation keys

* Translating cookie page

Updating from markdown template to html slim to follow pattern that we use acros the app.

* Removing section from default Welsh

Removing default error and helpers translation because it was overwriting normal buttons.

* Making sure that the url params are in url

When invalid form is submitted the "locale=en" was removed from URL. Now it's there.

* Fixing typo.

* Displaying feedback link based on language

The second feedback link points outside of the app and there is no Welsh version for it. So Cathie ask for the link to be same as the one in header when the language version is Welsh.

* Updating claim number translation in email

* Welsh address for office

If the office is Wales and the translation is Welsh the address is loaded from translation file. For the rest the address is loaded from DB which is in English.

* Adding missing mail translations

* Group claim template is now in welsh too

* Fixing missing translation in Welsh emails

* Fxing access denied for cy group claim template

The docker was ignoring the file so I'm saying not to ignore it so we can access it from web.

* Forget to update gitignore file

* Disabling language switch for active admin

Seting default locale value to en and disabling locale param for links.

* Adding Welsh pdf version.

* Fixing apostrophe in Welsh template

* Fixing links and translation in Welsh emails

* Fixing failing test

* Refactoring countable js

The lint validator was complainin about the length of the init method. I removed my addition to separate method.

* Fixing failing tests after locale moved to url

When I moved the locale from params to be part of the URL some test failed. This is just the fix.

* Fixing the layout for error message on employment page

* Trying to fix failing feature test

* Updating uppercase for "submit form"

* Removing extra space from end of line

* Reverting the flatten config change.

* Updating Welsh pdf template

* Fixin link in access email

* Fixing spacing on employment page

* Fixing checkbox size in welsh pdf

* Changed URL for refund profile selection page - now includes language

* Changed URL in all page objects to include /en in url

* Fixing error message spacing on claimant page

* Fixing additional padding issue

* Adding missing translations

* Adding json format responses for the errors

* Adding the missing html format respons